Move LuaGetLastError and exceptions to separate unit, update make files
* Include <stdexcept> which seemed to be missing
* Can possibly fix GCC and Clang errors
