+++ LuaCppInterface/LuaCppInterface.vcxproj.filters
@@ -33,6 +33,9 @@
     <ClCompile Include="luastringconversion.cpp">
       <Filter>Source Files</Filter>
     </ClCompile>
+    <ClCompile Include="luaerror.cpp">
+      <Filter>Source Files</Filter>
+    </ClCompile>
   </ItemGroup>
   <ItemGroup>
     <ClInclude Include="luacppinterface.h">
@@ -84,6 +87,9 @@
       <Filter>Header Files</Filter>
     </ClInclude>
     <ClInclude Include="luastringconversion.h">
+      <Filter>Header Files</Filter>
+    </ClInclude>
+    <ClInclude Include="luaerror.h">
       <Filter>Header Files</Filter>
     </ClInclude>
   </ItemGroup>
