+++ LuaCppInterface/Makefile.am
@@ -3,5 +3,5 @@ if COND_GCOV
 endif
 
 lib_LIBRARIES = libluacppinterface.a
-libluacppinterface_a_SOURCES = luacoroutine.cpp luacppinterface.cpp luafunction.cpp luareference.cpp luatable.cpp
+libluacppinterface_a_SOURCES = luacoroutine.cpp luacppinterface.cpp luafunction.cpp luareference.cpp luatable.cpp luaerror.cpp luastringconversion.cpp 
 AM_CXXFLAGS = $(MAYBE_COVERAGE) -I../lua/src $(BOOST_CPPFLAGS)
