+++ LuaCppInterface/luacppinclude.h
@@ -1,7 +1,12 @@
 #pragma once
+#ifndef LUACPPINCLUDE
+#define LUACPPINCLUDE
+
 
 extern "C++" {
 #include "lua.h"
 #include "lualib.h"
 #include "lauxlib.h"
 }
+
+#endif // LUACPPINCLUDE
